blkio: Increment the blkio cgroup stats for real now
Patchwork-id: 24657
O-Subject: [PATCH 05/19] blkio: Increment the blkio cgroup stats for real now
Bugzilla: 586182
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

From: Divyesh Shah <dpshah@google.com>

BZ:58612

commit 9195291e5f05e01d67f9a09c756b8aca8f009089
Author: Divyesh Shah <dpshah@google.com>

We also add start_time_ns and io_start_time_ns fields to struct request
here to record the time when a request is created and when it is
dispatched to device. We use ns uints here as ms and jiffies are
not very useful for non-rotational media.

Signed-off-by: Divyesh Shah<dpshah@google.com>
Signed-off-by: Jens Axboe <jens.axboe@oracle.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>