expose the statistics in blkio.time and blkio.sectors for the root cgroup
Patchwork-id: 24669
O-Subject: [PATCH 06/19] block: expose the statistics in blkio.time and
	blkio.sectors for the root cgroup
Bugzilla: 586182
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

From: Ricky Benitez <rickyb@google.com>

BZ:58612

commit a74b2adae06265b8cfa335d7d40d4a5abd11e977
Author: Ricky Benitez <rickyb@google.com>

Currently, the io statistics for the root cgroup are maintained, but
they are not shown because the device information is not available at
the point that the root blkio cgroup is created. This patch updates
the device information when the statistics are updated so that the
statistics become visible.

Signed-off-by: Ricky Benitez <rickyb@google.com>
Acked-by: Vivek Goyal <vgoyal@redhat.com>
Signed-off-by: Jens Axboe <jens.axboe@oracle.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>