blkio: Add io_queued and avg_queue_size stats
Patchwork-id: 24670
O-Subject: [PATCH 09/19] blkio: Add io_queued and avg_queue_size stats
Bugzilla: 586182
RH-Acked-by: Jeff Moyer <jmoyer@redhat.com>

From: Divyesh Shah <dpshah@google.com>

BZ:58612

commit cdc1184cf4a7bd99f5473a91244197accc49146b
Author: Divyesh Shah <dpshah@google.com>

These stats are useful for getting a feel for the queue depth of the cgroup,
i.e., how filled up its queues are at a given instant and over the existence of
the cgroup. This ability is useful when debugging problems in the wild as it
helps understand the application's IO pattern w/o having to read through the
userspace code (coz its tedious or just not available) or w/o the ability
to run blktrace (since you may not have root access and/or not want to disturb
performance).

Signed-off-by: Divyesh Shah<dpshah@google.com>
Signed-off-by: Jens Axboe <jens.axboe@oracle.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>