Reduce polling when interrupts are available
Patchwork-id: 24680
O-Subject: [PATCH RHEL6 5/8] ipmi: Reduce polling when interrupts are available
Bugzilla: 584106
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

If we're not currently in the middle of a transaction, and if we have
interrupts, there's no real reason to poll the controller more frequently
than the core IPMI code does. Set the interrupt_disabled flag appropriately
as the interrupt state changes, and make the timeout code reset itself
only if the transaction is incomplete or we have no interrupts.

Signed-off-by: Matthew Garrett <mjg@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>