Attempt to register multiple SIs of the same type
Patchwork-id: 24681
O-Subject: [PATCH RHEL6 7/8] ipmi: Attempt to register multiple SIs of the same
	type
Bugzilla: 584106
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Some odd systems may have multiple BMCs, and we want to be able to support
them. Let's make the assumption that if a system legitimately has multiple
BMCs then each BMC's SI will be of the same type, and also that we won't
see multiple SIs of the same type unless we have multiple BMCs. If these
hold true then we should register all SIs of the same type.

Signed-off-by: Matthew Garrett <mjg@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>