Change timeout and event poll to one second
Patchwork-id: 24683
O-Subject: [PATCH RHEL6 8/8] ipmi: Change timeout and event poll to one second
Bugzilla: 584106
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

From: Corey Minyard <cminyard@mvista.com>

The timeouts in IPMI are in the 1-5 second range in message handling,
so a 1 second timeout is a reasonable thing to do.  This should help
with reducing power consumption on idle systems.

Signed-off-by: Corey Minyard <cminyard@mvista.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>