Fall back to manually changing SCI_EN
Patchwork-id: 24687
O-Subject: [RHEL6 PATCH] acpi: Fall back to manually changing SCI_EN
Bugzilla: 587008
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

bz: 587008

The ACPI spec tells us that the ACPI SCI_EN bit is under hardware control
and shouldn't be touched by the OS. It seems that the Leading Other OS
ignores this and some machines expect this behaviour. We have a blacklist
for these, but given that we're able to detect the failure case and the
alternative to breaking the spec is letting the machine crash and burn,
let's try falling back when we know the alternative is a mostly-dead
machine.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>