nmi_watchdog: use __cpuinit for 32-bit nmi_watchdog_default
Patchwork-id: 24690
O-Subject: [RHEL6 PATCH]: nmi_watchdog: use __cpuinit for 32-bit
	nmi_watchdog_default
Bugzilla: 586967
RH-Acked-by: Pete Zaitcev <zaitcev@redhat.com>
RH-Acked-by: Bob Picco <bpicco@redhat.com>
RH-Acked-by: Dean Nelson <dnelson@redhat.com>

[Aside: I've noticed recently that patches generated by bugzilla are adding
extra tabs at the end of the ---/+++ lines in patches.  AFAICT these extra
lines are not causing problems with patch or diffstat.  However, just to be
safe -- Aris could you confirm that this patch format is okay?  If it
is not I think we'd all like to know so we can <scratch>yell at</scratch>
calmly explain to the bugzilla owners that we have a problem. ... Thanks.]

Repeated soft online and offline of cpus results in a panic.  The panic is
due to a nmi_watchdog_default() being declared __init.

Effects 32-bit x86 only.  64-bit was resolved in BZ 567601.

Successfully tested by me on an AMD Dinar and an Intel Nehalem-EX system, AMD
systems by Bhavna and Frank @ AMD.

Resolves BZ 586967.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>