zcore: Fix reipl device detection
Patchwork-id: 24696
O-Subject: [RHEL6 PATCH 1/1] [s390] zcore: Fix reipl device detection
Bugzilla: 587025
RH-Acked-by: Pete Zaitcev <zaitcev@redhat.com>

Description
-----------
zfcpdump fails and no kernel dump is created:
The reipl device information is passed from the kernel to zfcpdump
using a pointer in the lowcore (0xe00) that points to the reipl
information.  Currently, if that pointer is not zero, we copy the
reipl information.  If the pointer is not initialized and points
outside the accessible memory, it can happen that the memory copy
fails.  In that case we currently stop the initialization of zcore
which leads to a failing kernel dump.

The solution is to disable reipl after dump and continue with zcore
intialization.

Bugzilla
--------
BZ 587025
https://bugzilla.redhat.com/show_bug.cgi?id=587025

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.34
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=76ef964c78797f9baed7c2f9a58f696e86d8a048

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.