Fetch valid frequencies for powernow_k8.o from ACPI _PST table
Patchwork-id: 24736
O-Subject: Re: [RHEL6 PATCH REPOST] Fetch valid frequencies for powernow_k8.o
	from	ACPI _PST table
Bugzilla: 464630
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

Commit-ID:  679370641e3675633cad222449262abbe93a4a2a
Gitweb:
http://git.kernel.org/tip/679370641e3675633cad222449262abbe93a4a2a

1) The patch I submitted (see attached) has the necessary spaces just
like the upstream patch.
It appears that it got messed up during the SMTP transit.   It looks
fine in both my original
post and repost - checked both emails in RHML.
2) Look at the else statement in patch.
If processor is family 0x11, or family 0x10 model 0x9 or lower, then get
the values from the MSR.   Else get from ACPI.   I believe that this
logic covers all future processors.

Bhavna

Resolves BZ 464630

Core Performance Boost allows a CPU to run a core at speed higher than normal.
What happens is when say three core of a quad-core CPU are in a low power
state than the one core runnning can use additional power and thermal resource
from the other idle cores to go into a faster p-state than when all processors
are running.  This functionality is handled in hardware by specially create
P-states and needs no software support to enable the feature.

However through testing on a Pharoah Hound system, I found that the available
frequencies reported is the boost frequency.  For example, on a 2.8GHz P0,
the reported P0 is 3.3GHz, which is the boost frequency. This is because
the powernow-k8 driver gets the frequencies from the HW pstate registers
(which are more accurate), instead of the ACPI _PSS objects.  With CPB, the
reported valid frequencies need to be fetched from  ACPI.

Tested successfully on family 0x10, model 10, Pharoah Hound system.
The patch is in the -mm tree, and AMD highly recommends that the patch
be applied.

Please review and let me know if you see any issues.  Thanks.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>