KVM: fix the handling of dirty bitmaps to avoid overflows
Patchwork-id: 24784
O-Subject: [RHEL6 KVM 6/7] KVM: fix the handling of dirty bitmaps to avoid
	overflows
Bugzilla: 588811
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Gleb Natapov <gleb@redhat.com>
RH-Acked-by: Avi Kivity <avi@redhat.com>

commit 87bf6e7de1134f48681fd2ce4b7c1ec45458cb6d
From: Takuya Yoshikawa <yoshikawa.takuya@oss.ntt.co.jp>

KVM: fix the handling of dirty bitmaps to avoid overflows

Int is not long enough to store the size of a dirty bitmap.

This patch fixes this problem with the introduction of a wrapper
function to calculate the sizes of dirty bitmaps.

Note: in mark_page_dirty(), we have to consider the fact that
  __set_bit() takes the offset as int, not long.

Signed-off-by: Takuya Yoshikawa <yoshikawa.takuya@oss.ntt.co.jp>
Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>
Bugzilla: 588811

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>