nmi watchdog: use generic interrupt source to determine deadlocks
Patchwork-id: 24898
O-Subject: [RHEL-6 PATCH v3] [x86_64] nmi watchdog: use generic interrupt
	source to determine deadlocks
Bugzilla: 574570
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=574570

When the nmi watchdog was originally written it used the local apic
interrupts as the mechanism to determine if a cpu was deadlock or not (if
it
stopped incrementing the cpu was toast).

Somewhere in 2.6.2X the clock infrastructure was re-written to no longer
have TSC as the primary clock.  Instead HPET became the default if
available.  As a result, on those machines that used HPET, the local
interrupts stat stopped incrementing and instead the HPET stat
incremented.

The nmi watchdog was not updated to account for this and therefore thought
a
perfectly fine running system was deadlocking because its local interrupts
were not updating.  Depending on how much work a cpu was given, determined
how quickly it would fall over.  Therefore some people saw it during
install, others randomly later (and probably ignore it as an unstable
RHEL-6
kernel :-) ).

This patch implements an hrtimer that fires 'nmi_hz' times a second (most
of
the time once).  The idea is that if hrtimer is not firing than interrupts
are not working and the cpu is probably deadlocked.

This is the generic approach idea upstream came up with to avoid the mess
of
dealing with different clocks.  Unfortunately, I am in the middle of
rewriting the nmi watchdog code, so this change will make it upstream a
little bit differently.  But the approach is the same.

Tested on various boxes that had issues before and everything works fine
now.

V2:
The IO_APIC path was broken and caused the hrtimer to be kicked off twice
with the same variable leading to fun things.  Tested on RHEL-6 FV on
RHEL-5 Xen, where the original problem was found.  Also tested on bare
metal with nmi_watchdog=1.  Everything works fine now.

V3:
Make Prarit happy.

Please ACK.

Signed-off-by: Don Zickus <dzickus@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>