gfs2: fix oops while copying from ext3 to gfs2
Patchwork-id: 24988
O-Subject: [RHEL6 PATCH][GFS2] Bug 586009 - CVE-2010-1436 kernel: oops while
	cp from ext3 to gfs2 [rhel-6.0]
Bugzilla: 586009
CVE: CVE-2010-1436
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>

This is the upstream/rhel6 fix for this bug. This patch differs
from the RHEL5 fix (555754) which simply writes to the 8-byte
value field of the quota. In upstream/rhel6 quota code, we're
required to write the entire quota (88 bytes) which can be split
across a page boundary. We check for such quotas, and read/write
the two parts from/to the corresponding pages holding these parts.

With this patch, I don't see the bug anymore using the reproducer
in bz 555754. I successfully ran a couple of simple tests/mounts/
umounts and it doesn't seem like this patch breaks anything else.

Resolves: rhbz#586009
Signed-off-by: Abhi Das <adas@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>