sched: Make tunable scaling style configurable
Patchwork-id: 25008
O-Subject: [RHEL6 PATCH 2/3] sched: Make tunable scaling style configurable
Bugzilla: 590748
RH-Acked-by: Jiri Olsa <jolsa@redhat.com>

Description
-----------
As scaling now takes place on all kind of CPU add/remove events a user
that configures values via proc should be able to configure if his set
values are still rescaled or kept whatever happens.

As the comments state that log2 was just a second guess that worked the
interface is not just designed for on/off, but to choose a scaling type.
Currently this allows none, log, and linear, but more important it allows
us to keep the interface even if someone has an even better idea how to
scale the values.

Bugzilla
--------
BZ 590748
https://bugzilla.redhat.com/show_bug.cgi?id=590748

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.33
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=1983a922a1bc843806b9a36cf3a370b242783140

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>