hpsa: update to 2.0.2
Patchwork-id: 25040
O-Subject: [RHEL6 PATCH] hpsa: update to 2.0.2
Bugzilla: 587418
RH-Acked-by: Rob Evers <revers@redhat.com>

This is for bz#587418

This patch comes from HP. hpsa is a new
driver which should be used in the future instead
of cciss for the newer(not older than 4 years)
HP controllers. The older cciss will still be
used for the older cards.

The upstream version of the  whole driver in scsi-misc tree,
is almost the same the diferencies are minor .

I've tested this lightly on x86_64 with a P410i controller
The main testing and support will come from HP.

Tomas

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>