CacheFiles: Fix error handling in cachefiles_determine_cache_security()
Patchwork-id: 25062
O-Subject: [PATCH 9/9] CacheFiles: Fix error handling in
	cachefiles_determine_cache_security()
Bugzilla: 591894
RH-Acked-by: Steve Dickson <SteveD@redhat.com>

cachefiles_determine_cache_security() is expected to return with a
security override in place.  However, if set_create_files_as() fails, we
fail to do this.  In this case, we should just reinstate the security
override that was set by the caller.

Furthermore, if set_create_files_as() fails, we should dispose of the
new credentials we were in the process of creating.

Signed-off-by: David Howells <dhowells@redhat.com>
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>