export paravirtual cpuid flags in KVM_GET_SUPPORTED_CPUID
Patchwork-id: 25096
O-Subject: [PATCH RHEL6.0 3/4] export paravirtual cpuid flags in
	KVM_GET_SUPPORTED_CPUID
Bugzilla: 592296
RH-Acked-by: Rik van Riel <riel@redhat.com>

RH-Author: Glauber Costa <glommer@redhat.com>
RH-Upstream-status: kvm/master
RH-Bugzilla: 592296

Right now, we were using individual KVM_CAP entities to communicate
userspace about which cpuids we support. This is suboptimal, since it
generates a delay between the feature arriving in the host, and
being available at the guest.

A much better mechanism is to list para features in KVM_GET_SUPPORTED_CPUID.
This makes userspace automatically aware of what we provide. And if we
ever add a new cpuid bit in the future, we have to do that again,
which create some complexity and delay in feature adoption.

Signed-off-by: Glauber Costa <glommer@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>