Try using new kvm clock msrs
Patchwork-id: 25101
O-Subject: [PATCH RHEL6.0 3/4] Try using new kvm clock msrs
Bugzilla: 569603
RH-Acked-by: Rik van Riel <riel@redhat.com>

RH-Author: Glauber Costa <glommer@redhat.com>
RH-Upstream-status: kvm/master
RH-Bugzilla: 569603

We now added a new set of clock-related msrs in replacement of the old
ones. In theory, we could just try to use them and get a return value
indicating they do not exist, due to our use of kvm_write_msr_save.

However, kvm clock registration happens very early, and if we ever
try to write to a non-existant MSR, we raise a lethal #GP, since our
idt handlers are not in place yet.

So this patch tests for a cpuid feature exported by the host to
decide which set of msrs are supported.

Signed-off-by: Glauber Costa <glommer@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>