Fix a number of module signing bugs
Patchwork-id: 25091
O-Subject: [RHEL6 PATCH] MODSIGN: Fix module signing
Bugzilla: 581965
RH-Acked-by: Roland McGrath <roland@redhat.com>

Fix the module signing code in a number of ways [BZ 581965]:

 (1) Use __stringify() to generate the ELF note name, thus dealing the weird
     behaviour of the '#' cpp op.

 (2) Make use of the canonicalised section table to order the input of data
     into the digest in the kernel and in mod-extract.

 (3) Ignore gcc's build ID note as this has issues with the debugedit program
     (hopefully to be fixed by BZ 590947, in which case we can eliminate the
     skip).

 (4) Include the headers of BSS sections in the digest.

Also included are some changes to make future debugging easier without the need
to recompile the kernel:

 (5) Make it possible to turn on module signature check debugging dynamically,
     including on the kernel command line.

 (6) Print a little more information when it does go wrong.

And a cleanup for some warnings:

 (7) Eliminate warnings in mod-extract.c from the byteorder functions.

v2:
The use of sg_init_one() to load vmalloc'd memory into a scatter-gather list
does not work [BZ 599266].

Alter the module signing code to use the crypto shash interface rather than
the crypto hash interface so that we can pass in buffer pointers directly
rather than having to use a scatter-gather list.

Further, to aid debugging, if ksign_debug=1 is added to the kernel command line
(or the appropriate sysfs file set to 1), the intermediate SHA1 sum, consisting
of a digest just of the canonicalised file contents, will be dumped to dmesg:

	Modsign digest: 74e85a491150b62557e00f1326adc28ed7925dc1

This can then be compared to the .ko.digest file in the build tree:

	warthog>sha1sum fs/ext2/ext2.ko.digest
	74e85a491150b62557e00f1326adc28ed7925dc1  fs/ext2/ext2.ko.digest

This has been tested on both an i686 and an x86_64 installation by (a) booting,
and then (b) loading the ext2 module.

Signed-off-by: David Howells <dhowells@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>