audit: dynamically allocate audit_names when not enough space is in the names array
Patchwork-id: 25118
O-Subject: [PATCH RHEL6] audit: dynamically allocate audit_names when not enough
	space is in the names array
Bugzilla: 586108
RH-Acked-by: Josef Bacik <josef@redhat.com>

BZ 586108

This patch does 2 things.  First it reduces the number of audit_names
allocated in every audit context from 20 to 5.  5 should be enough for all
'normal' syscalls (rename being the worst).  Some syscalls can still touch
more the 5 inodes such as mount.  When rpc filesystem is mounted it will
create inodes and those can exceed 5.  To handle that problem this patch will
dynamically allocate audit_names if it needs more than 5.  This should
decrease the typicall memory usage while still supporting all the possible
kernel operations.

Not upstream as Al would like to try to address the problem a different way.
He would like to try to find all of the places that generate numerous inode
records in a single syscall and shut those up.  I'm not convinced that's the
best way, but this is clearly the best way for RHEL6 since that it going to
take some larger upstream changes.

Signed-off-by: Eric Paris <eparis@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>