only call start-cpu when a CPU is stopped
Patchwork-id: 25129
O-Subject: [PATCH RHEL6.0 BZ592440 2/2] powerpc: only call start-cpu when a CPU
	is stopped
Bugzilla: 592440
RH-Acked-by: David Howells <dhowells@redhat.com>
RH-Acked-by: Neil Horman <nhorman@redhat.com>

RHBZ#:
======
https://bugzilla.redhat.com/show_bug.cgi?id=592440

Description:
============
Patch 2/2: powerpc: only call start-cpu when a CPU is stopped

Commit id:
http://git.kernel.org/?p=linux/kernel/git/benh/powerpc.git;a=commit;h=aef40e87d866355ffd279ab21021de733242d0d5

Currently we always call start-cpu irrespective of if the CPU is
stopped or not. Unfortunatley on POWER7, firmware seems to not like
start-cpu being called when a cpu already been started.  This was not
the case on POWER6 and earlier.

This patch checks to see if the CPU is stopped or not via an
query-cpu-stopped-state call, and only calls start-cpu on CPUs which
are stopped.

This fixes a bug with kexec on POWER7 on PHYP where only the primary
thread would make it to the second kernel.

Reported-by: Ankita Garg <ankita@linux.vnet.ibm.com>
Cc: <stable@kernel.org>
Signed-off-by: Michael Neuling <mikey@neuling.org>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>

Brew:
=====
Built on all platforms. All patches brewed together.
https://brewweb.devel.redhat.com/taskinfo?taskID=2446727

===============================================================
Steve Best
IBM on-site partner

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>