quota: use flags interface for dquot alloc/free space
Patchwork-id: 25136
O-Subject: [PATCH RHEL6 3/5] quota: use flags interface for dquot alloc/free
	space
Bugzilla: 587095
RH-Acked-by: Josef Bacik <josef@redhat.com>

This is for bug 587095

It's not upstream yet, but it's ack'd by Jan Kara and in the
ext4 patch queue.

Switch __dquot_alloc_space and __dquot_free_space to take flags
to indicate whether to warn and/or to reserve (or free reserve).

This is slightly more readable at the callpoints, and makes it
cleaner to add a "nofail" option in the next patch.

Signed-off-by: Eric Sandeen <sandeen@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>