quota: add the option to not fail with EDQUOT in block allocation
Patchwork-id: 25137
O-Subject: [PATCH RHEL6 4/5] quota: add the option to not fail with EDQUOT in
	block allocation
Bugzilla: 587095
RH-Acked-by: Josef Bacik <josef@redhat.com>

This is for bug 587095

It's not upstream yet, but it's ack'd by Jan Kara and in the
ext4 patch queue.

To simplify metadata tracking for delalloc writes, ext4
will simply claim metadata blocks at allocation time, without
first speculatively reserving the worst case and then freeing
what was not used.

To do this, we need a mechanism to track allocations in
the quota subsystem, but potentially allow that allocation
to actually go over quota.

This patch adds a DQUOT_SPACE_NOFAIL flag and function
variants for this purpose.

Signed-off-by: Eric Sandeen <sandeen@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>